% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{blue_shades}
\alias{blue_shades}
\title{Blue Shades}
\usage{
blue_shades(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing blue shades.
}
\description{
This function provides a range of blue shades used in TRC visualizations.
}
\examples{
blue_shades()      # Default order
blue_shades(TRUE)  # Reversed order
}
