% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{orange_shades}
\alias{orange_shades}
\title{Orange Shades}
\usage{
orange_shades(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing orange shades.
}
\description{
This function provides a range of orange shades used in TRC visualizations.
}
\examples{
orange_shades()      # Default order
orange_shades(TRUE)  # Reversed order
}
