% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{purple_shades}
\alias{purple_shades}
\title{Purple Shades}
\usage{
purple_shades(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing purple shades.
}
\description{
This function provides a range of purple shades used in TRC visualizations.
}
\examples{
purple_shades()      # Default order
purple_shades(TRUE)  # Reversed order
}
