% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{red_shades}
\alias{red_shades}
\title{Red Shades}
\usage{
red_shades(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing red shades.
}
\description{
This function provides a range of red shades used in TRC visualizations.
}
\examples{
red_shades()      # Default order
red_shades(TRUE)  # Reversed order
}
