% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{teal_shades}
\alias{teal_shades}
\title{Teal Shades}
\usage{
teal_shades(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing teal shades.
}
\description{
This function provides a range of teal shades used in TRC visualizations.
}
\examples{
teal_shades()      # Default order
teal_shades(TRUE)  # Reversed order
}
