% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{trc_main_colours}
\alias{trc_main_colours}
\title{Main TRC Colors}
\usage{
trc_main_colours(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing the main colors.
}
\description{
This function provides the main color palette used in TRC visualizations.
}
\examples{
trc_main_colours()      # Default order
trc_main_colours(TRUE)  # Reversed order
}
