% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{trc_oranges}
\alias{trc_oranges}
\title{Orange Shades}
\usage{
trc_oranges(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing orange shades.
}
\description{
This function provides a range of orange shades used in TRC visualizations.
}
\examples{
trc_oranges()      # Default order
trc_oranges(TRUE)  # Reversed order
}
