% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{trc_purples}
\alias{trc_purples}
\title{Purple Shades}
\usage{
trc_purples(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing purple shades.
}
\description{
This function provides a range of purple shades used in TRC visualizations.
}
\examples{
trc_purples()      # Default order
trc_purples(TRUE)  # Reversed order
}
