% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{trc_teals}
\alias{trc_teals}
\title{Teal Shades}
\usage{
trc_teals(reverse = FALSE)
}
\arguments{
\item{reverse}{Logical. If TRUE, reverses the order of the colors.}
}
\value{
A character vector of hex color codes representing teal shades.
}
\description{
This function provides a range of teal shades used in TRC visualizations.
}
\examples{
trc_teals()      # Default order
trc_teals(TRUE)  # Reversed order
}
